Remove unnecessary extension from imports
