+++ src/components/extra-films-list.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from './abstract-component.js';
+import {AbstractComponent} from './abstract-component';
 
 export class ExtraFilmsContainer extends AbstractComponent {
   constructor(title = ``) {
