+++ src/components/filters-stats.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from './abstract-component.js';
+import {AbstractComponent} from './abstract-component';
 
 export class Filters extends AbstractComponent {
   constructor(films = []) {
