+++ src/components/show-more-btn.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from './abstract-component.js';
+import {AbstractComponent} from './abstract-component';
 
 export class ShowMoreBtn extends AbstractComponent {
   constructor() {
