--- src/utils.js
@@ -1,7 +1,0 @@
-export const htmlStringToElement = (html) => {
-  const template = document.createElement(`template`);
-  const trimmedHtml = html.trim();
-  template.innerHTML = trimmedHtml;
-
-  return template.content.firstChild;
-};
